fix: Updates target framework to .NET 9
Updates the target framework to .NET 9.
This change ensures compatibility with the latest .NET runtime and incorporates the latest features and improvements.

Also updates the package dependencies to their respective latest versions to align with the .NET 9 upgrade.
Updates the integration tests to target new coordinates and also adds a new test case for retrieving weather for the next day.